Add support to log in using site name
In the tydom app, users tend to refet to their homes with a friendly
name.

Currently, we require the full mac address which is not trivial to get.

With this change, we will be able to provide the site name and perform a
login using the site name.
This contributes to improving the user experience configuring the
integraiton for the first time